Workaround spec demod bin count parsing bug
I'm seeing occasional spec demod data with inconsistent bin count and bin
data, but only when the poll request was for less than 1620 bins. This
may or may not be an old bug. As a workaround, the parser will now abort
and return 0 bins if the bin count doesn't match the bin data length.